feat: move api from util to lib folder
