+++ src/app/dashboard/businessOption/page.tsx
@@ -4,7 +4,7 @@ import React, { useEffect, useState } from "react";
 
 import LogoutBtn from "@/components/shared/LogoutBtn";
 import TradeMark from "@/components/shared/TradeMark";
-import getUserBusiness from "@/utils/getUserBusinesses";
+import getUserBusiness from "@/lib/getUserBusinesses";
 import styles from "./businessOption.module.css";
 import { useBusinessContext } from "@/context/BusinessContext";
 import { useRouter } from "next/navigation";
